Create useSendMessage and useFileValidation custom hooks
